extract non-generated part of MeshStrings file to extenstions file
